Add feature to pause auto play when user input is detected
Added a variable `pauseOnUserImput` which takes a `Duration`.
If `autoPlay` is set to true and `pauseOnUserInput` is not null, the
`PageView` will be wrapped in a `GestureDetector` listening for
pan gestures.

Also did some minor consistency changes,
i.e. Hard typed a few untyped values and removed the 'new' keywords.
Almost all values were already hard typed and 'new' was used in
only a few places.
Now each line of code follows the same code convention.

My linting options were different from the projects, so I did my best
to follow the repo manually.
Feel free to auto format the code with the previously used setting!

Premise:
When using a sliding carousel you might want the
user to be able to interact with a slide.
For now, the user can swipe back if he/she see something they
like, but the carousel will keep iterating through its children.
This forces the user to constantly swipe back to the page they
were interacting with.
For example, imagen a commercial display rolling adds in a
waiting room. If something interest a person he or she might
want to pause the slide to read all the information.

This feature lets the developer implement a way too pause the
auto play feature for a set amount of time, resetting each time
new input is detected.
This way the user can interact with the application and after they
leave, it will go back to circling the widgets.