Deprecate old variables for backwards compatibility
Marked the changed properties as deprecated and introduced the new.
Made assertions to check if both are being used at the same time and
passed default values when both are null.

Added a short explanation of the new feature to readme under the FAQ section.

Updated file structure to follow flutters style guide.
Most noticeably changed the order to match the guidelines:
0. Constructors, with the default constructor first.
1. Constants of the same type as the class.
2. Static methods that return the same type as the class.
3. Final fields that are set from the constructor.
4. Other static methods.
5. Static properties and constants.
6. Mutable properties, each in the order getter, private field, setter, without newlines separating them.
7. Read-only properties (other than hashCode).
8. Operators (other than ==).
9. Methods (other than toString and build).
10. The build method, for Widget and State classes.
11. operator ==, hashCode, toString, and diagnostics-related methods, in that order.

Reference:
https://github.com/flutter/flutter/wiki/Style-guide-for-Flutter-repo

Squashed—————

Updated readme to reflect the new changes.
Also added three .gif files and replaced the outdated one.
These Gifs show the apps functionality.

Relocated dot indicators to below the carousel since they were hard
to see on top of darker images.

Added documentation to all input params and reordered them so
that they come in a grouped order.
That is, items first, then layout params followed by auto play features.

I also changed  a few argument names that confused me in the beginning.
For example I changed  "distortion"  to "enlargeCenterPage".
Perhaps a little bit clumsy, but straight forward and easy to understand if
you just started using the lib.

Feel free to revert any changes you feel is beneficial for the lib.

Added a PtP (Push to pause) widget at the bottom of
the demo app.

Remodeled the app for readability.
Collected all widgets at top of the build method and
then referred to them in the returning widget.
It makes it easier for adopting developers to find the
example they are looking for, and its scaleable.

Added labels above each example describing the
property on display.