Add feature to disable infinity scroll
By popular demand I created a simple bool, `loopItems`,
when set to false will reduce the item count to items.count.

This results in a static length page viewer that still retains
all the other benefits of this library (animations, buttons, autoplay etc...).

I also added comments to this and the vertical scroll feature i rebased against.
I realized now that it still have a PR open, but I added some info about it in the
Readme and change-log.

Bumped version to 1.2.0